refactor otel into middleware
