+++ src/middleware/index.ts
@@ -5,5 +5,13 @@
 
 export { MiddlewareEngine, MiddlewareException } from './MiddlewareEngine'
 export { ProductCheck, type ProductCheckOptions } from './ProductCheck'
+export {
+  OpenTelemetryMiddleware,
+  type OpenTelemetryOptions,
+  SEARCH_LIKE_ENDPOINTS,
+  STRING_QUERY_ENDPOINTS,
+  NDJSON_ENDPOINTS,
+  SEARCH_QUERY_MAX_LENGTH
+} from './OpenTelemetry'
 export { MiddlewareName, MiddlewarePriority } from './types'
 export type { Middleware, MiddlewareContext, MiddlewareResult } from './types'
