+++ src/symbols.ts
@@ -34,6 +34,4 @@ export const kNdjsonContentType = Symbol('ndjson content type')
 export const kAcceptHeader = Symbol('accept header')
 export const kRedaction = Symbol('redaction')
 export const kRetryBackoff = Symbol('retry backoff')
-export const kOtelTracer = Symbol('opentelemetry tracer')
-export const kOtelOptions = Symbol('opentelemetry options')
 export const kMiddlewareEngine = Symbol('middleware engine')
